+++ com.unity.ml-agents.extensions/Runtime/Input/IInputActionAssetProvider.cs
@@ -1,4 +1,4 @@
-#if MLA_INPUT_SYSTEM && UNITY_2019_4_OR_NEWER
+#if MLA_INPUT_SYSTEM
 using UnityEngine.InputSystem;
 
 namespace Unity.MLAgents.Extensions.Input
@@ -24,4 +24,4 @@ namespace Unity.MLAgents.Extensions.Input
         (InputActionAsset, IInputActionCollection2) GetInputActionAsset();
     }
 }
-#endif // MLA_INPUT_SYSTEM && UNITY_2019_4_OR_NEWER
+#endif // MLA_INPUT_SYSTEM