+++ com.unity.ml-agents/Runtime/Sensors/CameraSensorComponent.cs
@@ -76,6 +76,18 @@ namespace Unity.MLAgents.Sensors
         }
 
         [HideInInspector, SerializeField]
+        ObservationType m_ObservationType;
+
+        /// <summary>
+        /// The type of the observation.
+        /// </summary>
+        public ObservationType ObservationType
+        {
+            get { return m_ObservationType; }
+            set { m_ObservationType = value; UpdateSensor(); }
+        }
+
+        [HideInInspector, SerializeField]
         [Range(1, 50)]
         [Tooltip("Number of camera frames that will be stacked before being fed to the neural network.")]
         int m_ObservationStacks = 1;
@@ -106,30 +118,15 @@ namespace Unity.MLAgents.Sensors
         /// Creates the <see cref="CameraSensor"/>
         /// </summary>
         /// <returns>The created <see cref="CameraSensor"/> object for this component.</returns>
-        public override ISensor CreateSensor()
+        public override ISensor[] CreateSensors()
         {
-            m_Sensor = new CameraSensor(m_Camera, m_Width, m_Height, Grayscale, m_SensorName, m_Compression);
+            m_Sensor = new CameraSensor(m_Camera, m_Width, m_Height, Grayscale, m_SensorName, m_Compression, m_ObservationType);
 
             if (ObservationStacks != 1)
             {
-                return new StackingSensor(m_Sensor, ObservationStacks);
+                return new ISensor[] { new StackingSensor(m_Sensor, ObservationStacks) };
             }
-            return m_Sensor;
-        }
-
-        /// <summary>
-        /// Computes the observation shape of the sensor.
-        /// </summary>
-        /// <returns>The observation shape of the associated <see cref="CameraSensor"/> object.</returns>
-        public override int[] GetObservationShape()
-        {
-            var stacks = ObservationStacks > 1 ? ObservationStacks : 1;
-            var cameraSensorshape = CameraSensor.GenerateShape(m_Width, m_Height, Grayscale);
-            if (stacks > 1)
-            {
-                cameraSensorshape[cameraSensorshape.Length - 1] *= stacks;
-            }
-            return cameraSensorshape;
+            return new ISensor[] { m_Sensor };
         }
 
         /// <summary>