+++ com.unity.ml-agents/Runtime/Utilities.cs
@@ -1,10 +1,10 @@
 using System;
+using System.Diagnostics;
 
 namespace Unity.MLAgents
 {
     internal static class Utilities
     {
-
         /// <summary>
         /// Calculates the cumulative sum of an integer array. The result array will be one element
         /// larger than the input array since it has a padded 0 at the beginning.
@@ -26,10 +26,9 @@ namespace Unity.MLAgents
             return result;
         }
 
-#if DEBUG
+        [Conditional("DEBUG")]
         internal static void DebugCheckNanAndInfinity(float value, string valueCategory, string caller)
         {
-
             if (float.IsNaN(value))
             {
                 throw new ArgumentException($"NaN {valueCategory} passed to {caller}.");
@@ -39,7 +38,5 @@ namespace Unity.MLAgents
                 throw new ArgumentException($"Inifinity {valueCategory} passed to {caller}.");
             }
         }
-#endif
     }
-
 }