+++ com.unity.ml-agents/Tests/Editor/Inference/ParameterLoaderTest.cs
@@ -14,17 +14,12 @@ namespace Unity.MLAgents.Tests
     {
         public ISensor Sensor;
 
-        public override ISensor CreateSensor()
+        public override ISensor[] CreateSensors()
         {
-            return Sensor;
-        }
-
-        public override int[] GetObservationShape()
-        {
-            var shape = Sensor.GetObservationSpec().Shape;
-            return new int[] { shape[0], shape[1], shape[2] };
+            return new ISensor[] { Sensor };
         }
     }
+
     public class Test3DSensor : ISensor, IBuiltInSensor
     {
         int m_Width;
@@ -64,9 +59,9 @@ namespace Unity.MLAgents.Tests
         public void Update() { }
         public void Reset() { }
 
-        public SensorCompressionType GetCompressionType()
+        public CompressionSpec GetCompressionSpec()
         {
-            return SensorCompressionType.None;
+            return CompressionSpec.Default();
         }
 
         public string GetName()
@@ -291,7 +286,13 @@ namespace Unity.MLAgents.Tests
 
             var errors = BarracudaModelParamLoader.CheckModel(
                 model, validBrainParameters,
-                new ISensor[] { new VectorSensor(8), sensor_21_20_3.CreateSensor(), sensor_20_22_3.CreateSensor() }, new ActuatorComponent[0]
+                new ISensor[]
+                {
+                    new VectorSensor(8),
+                    sensor_21_20_3.CreateSensors()[0],
+                    sensor_20_22_3.CreateSensors()[0]
+                },
+                new ActuatorComponent[0]
             );
             Assert.AreEqual(0, errors.Count()); // There should not be any errors
         }
@@ -305,7 +306,7 @@ namespace Unity.MLAgents.Tests
 
             var errors = BarracudaModelParamLoader.CheckModel(
                 model, validBrainParameters,
-                new ISensor[] { sensor_21_20_3.CreateSensor() }, new ActuatorComponent[0]
+                new ISensor[] { sensor_21_20_3.CreateSensors()[0] }, new ActuatorComponent[0]
             );
             Assert.AreEqual(0, errors.Count()); // There should not be any errors
         }
@@ -318,7 +319,10 @@ namespace Unity.MLAgents.Tests
 
             var errors = BarracudaModelParamLoader.CheckModel(
                 model, validBrainParameters,
-                new ISensor[] { new VectorSensor(validBrainParameters.VectorObservationSize) }, new ActuatorComponent[0]
+                new ISensor[]
+                {
+                    new VectorSensor(validBrainParameters.VectorObservationSize)
+                }, new ActuatorComponent[0]
             );
             Assert.AreEqual(0, errors.Count()); // There should not be any errors
         }
@@ -333,7 +337,12 @@ namespace Unity.MLAgents.Tests
             brainParameters.VectorObservationSize = 9; // Invalid observation
             var errors = BarracudaModelParamLoader.CheckModel(
                 model, brainParameters,
-                new ISensor[] { sensor_21_20_3.CreateSensor(), sensor_20_22_3.CreateSensor() }, new ActuatorComponent[0]
+                new ISensor[]
+                {
+                    sensor_21_20_3.CreateSensors()[0],
+                    sensor_20_22_3.CreateSensors()[0]
+                },
+                new ActuatorComponent[0]
             );
             Assert.Greater(errors.Count(), 0);
 
@@ -341,7 +350,12 @@ namespace Unity.MLAgents.Tests
             brainParameters.NumStackedVectorObservations = 2;// Invalid stacking
             errors = BarracudaModelParamLoader.CheckModel(
                 model, brainParameters,
-                new ISensor[] { sensor_21_20_3.CreateSensor(), sensor_20_22_3.CreateSensor() }, new ActuatorComponent[0]
+                new ISensor[]
+                {
+                    sensor_21_20_3.CreateSensors()[0],
+                    sensor_20_22_3.CreateSensors()[0]
+                },
+                new ActuatorComponent[0]
             );
             Assert.Greater(errors.Count(), 0);
         }
@@ -354,7 +368,13 @@ namespace Unity.MLAgents.Tests
 
             var brainParameters = GetDiscrete1vis0vec_2_3action_recurrModelBrainParameters();
             brainParameters.VectorObservationSize = 1; // Invalid observation
-            var errors = BarracudaModelParamLoader.CheckModel(model, brainParameters, new ISensor[] { sensor_21_20_3.CreateSensor() }, new ActuatorComponent[0]);
+            var errors = BarracudaModelParamLoader.CheckModel(
+                model, brainParameters, new ISensor[]
+                {
+                    sensor_21_20_3.CreateSensors()[0]
+                },
+                new ActuatorComponent[0]
+            );
             Assert.Greater(errors.Count(), 0);
         }
 
@@ -388,12 +408,26 @@ namespace Unity.MLAgents.Tests
 
             var brainParameters = GetContinuous2vis8vec2actionBrainParameters();
             brainParameters.ActionSpec = ActionSpec.MakeContinuous(3); // Invalid action
-            var errors = BarracudaModelParamLoader.CheckModel(model, brainParameters, new ISensor[] { sensor_21_20_3.CreateSensor(), sensor_20_22_3.CreateSensor() }, new ActuatorComponent[0]);
+            var errors = BarracudaModelParamLoader.CheckModel(
+                model, brainParameters, new ISensor[]
+                {
+                    sensor_21_20_3.CreateSensors()[0],
+                    sensor_20_22_3.CreateSensors()[0]
+                },
+                new ActuatorComponent[0]
+            );
             Assert.Greater(errors.Count(), 0);
 
             brainParameters = GetContinuous2vis8vec2actionBrainParameters();
             brainParameters.ActionSpec = ActionSpec.MakeDiscrete(3); // Invalid SpaceType
-            errors = BarracudaModelParamLoader.CheckModel(model, brainParameters, new ISensor[] { sensor_21_20_3.CreateSensor(), sensor_20_22_3.CreateSensor() }, new ActuatorComponent[0]);
+            errors = BarracudaModelParamLoader.CheckModel(
+                model, brainParameters, new ISensor[]
+                {
+                    sensor_21_20_3.CreateSensors()[0],
+                    sensor_20_22_3.CreateSensors()[0]
+                },
+                new ActuatorComponent[0]
+            );
             Assert.Greater(errors.Count(), 0);
         }
 
@@ -405,12 +439,21 @@ namespace Unity.MLAgents.Tests
 
             var brainParameters = GetDiscrete1vis0vec_2_3action_recurrModelBrainParameters();
             brainParameters.ActionSpec = ActionSpec.MakeDiscrete(3, 3); // Invalid action
-            var errors = BarracudaModelParamLoader.CheckModel(model, brainParameters, new ISensor[] { sensor_21_20_3.CreateSensor() }, new ActuatorComponent[0]);
+            var errors = BarracudaModelParamLoader.CheckModel(
+                model, brainParameters,
+                new ISensor[] { sensor_21_20_3.CreateSensors()[0] },
+                new ActuatorComponent[0]
+            );
             Assert.Greater(errors.Count(), 0);
 
             brainParameters = GetContinuous2vis8vec2actionBrainParameters();
             brainParameters.ActionSpec = ActionSpec.MakeContinuous(2); // Invalid SpaceType
-            errors = BarracudaModelParamLoader.CheckModel(model, brainParameters, new ISensor[] { sensor_21_20_3.CreateSensor() }, new ActuatorComponent[0]);
+            errors = BarracudaModelParamLoader.CheckModel(
+                model,
+                brainParameters,
+                new ISensor[] { sensor_21_20_3.CreateSensors()[0] },
+                new ActuatorComponent[0]
+            );
             Assert.Greater(errors.Count(), 0);
         }
 
@@ -421,12 +464,30 @@ namespace Unity.MLAgents.Tests
 
             var brainParameters = GetHybridBrainParameters();
             brainParameters.ActionSpec = new ActionSpec(3, new[] { 3 }); // Invalid discrete action size
-            var errors = BarracudaModelParamLoader.CheckModel(model, brainParameters, new ISensor[] { sensor_21_20_3.CreateSensor(), sensor_20_22_3.CreateSensor() }, new ActuatorComponent[0]);
+            var errors = BarracudaModelParamLoader.CheckModel(
+                model,
+                brainParameters,
+                new ISensor[]
+                {
+                    sensor_21_20_3.CreateSensors()[0],
+                    sensor_20_22_3.CreateSensors()[0]
+                },
+                new ActuatorComponent[0]
+            );
             Assert.Greater(errors.Count(), 0);
 
             brainParameters = GetContinuous2vis8vec2actionBrainParameters();
             brainParameters.ActionSpec = ActionSpec.MakeDiscrete(2); // Missing continuous action
-            errors = BarracudaModelParamLoader.CheckModel(model, brainParameters, new ISensor[] { sensor_21_20_3.CreateSensor(), sensor_20_22_3.CreateSensor() }, new ActuatorComponent[0]);
+            errors = BarracudaModelParamLoader.CheckModel(
+                model,
+                brainParameters,
+                new ISensor[]
+                {
+                    sensor_21_20_3.CreateSensors()[0],
+                    sensor_20_22_3.CreateSensors()[0]
+                },
+                new ActuatorComponent[0]
+            );
             Assert.Greater(errors.Count(), 0);
         }
 
@@ -434,7 +495,16 @@ namespace Unity.MLAgents.Tests
         public void TestCheckModelThrowsNoModel()
         {
             var brainParameters = GetContinuous2vis8vec2actionBrainParameters();
-            var errors = BarracudaModelParamLoader.CheckModel(null, brainParameters, new ISensor[] { sensor_21_20_3.CreateSensor(), sensor_20_22_3.CreateSensor() }, new ActuatorComponent[0]);
+            var errors = BarracudaModelParamLoader.CheckModel(
+                null,
+                brainParameters,
+                new ISensor[]
+                {
+                    sensor_21_20_3.CreateSensors()[0],
+                    sensor_20_22_3.CreateSensors()[0]
+                },
+                new ActuatorComponent[0]
+            );
             Assert.Greater(errors.Count(), 0);
         }
     }