+++ com.unity.ml-agents/Tests/Runtime/Sensor/FloatVisualSensorTests.cs
@@ -64,9 +64,9 @@ namespace Unity.MLAgents.Tests
         public void Update() { }
         public void Reset() { }
 
-        public SensorCompressionType GetCompressionType()
+        public CompressionSpec GetCompressionSpec()
         {
-            return SensorCompressionType.None;
+            return CompressionSpec.Default();
         }
     }