+++ com.unity.ml-agents/Tests/Runtime/Sensor/SensorShapeValidatorTests.cs
@@ -50,15 +50,14 @@ namespace Unity.MLAgents.Tests
         public void Update() { }
         public void Reset() { }
 
-        public SensorCompressionType GetCompressionType()
+        public CompressionSpec GetCompressionSpec()
         {
-            return SensorCompressionType.None;
+            return CompressionSpec.Default();
         }
     }
 
     public class SensorShapeValidatorTests
     {
-
         [Test]
         public void TestShapesAgree()
         {
@@ -87,6 +86,7 @@ namespace Unity.MLAgents.Tests
             LogAssert.Expect(LogType.Assert, "Number of Sensors must match. 2 != 3");
             validator.ValidateSensors(sensorList1);
         }
+
         [Test]
         public void TestDimensionMismatch()
         {