+++ ml-agents/mlagents/trainers/settings.py
@@ -92,6 +92,11 @@ class ScheduleType(Enum):
     LINEAR = "linear"
 
 
+class ConditioningType(Enum):
+    HYPER = "hyper"
+    NONE = "none"
+
+
 @attr.s(auto_attribs=True)
 class NetworkSettings:
     @attr.s
@@ -115,6 +120,7 @@ class NetworkSettings:
     num_layers: int = 2
     vis_encode_type: EncoderType = EncoderType.SIMPLE
     memory: Optional[MemorySettings] = None
+    goal_conditioning_type: ConditioningType = ConditioningType.HYPER
 
 
 @attr.s(auto_attribs=True)